flow fixes

+++ src/ethereum/contracts/HealthChain.sol
@@ -15,19 +15,19 @@ contract HealthChain {
         string ill_id;
         string name;
         address doctor;
-        uint created_at;
+        uint256 created_at;
     }
     
     struct Patient {
         string name;
-        uint age;
+        uint256 age;
         address wallet_address;
         string gender;
     }
     
     struct Doctor {
         string name;
-        uint age;
+        uint256 age;
         string gender;
         address wallet_address;
         string hospital_name;
@@ -38,7 +38,7 @@ contract HealthChain {
         string ill_id;
         string name;
         Doctor doctor;
-        uint created_at;
+        uint256 created_at;
     }
     mapping(address => Patient) public patients;//all patients
     mapping(address => Doctor) public doctors;//all doctors
@@ -61,7 +61,7 @@ contract HealthChain {
     function getDoctor() public view returns (Doctor memory) {
         return doctors[msg.sender];
     }
-    function createDoctor(string memory name , uint age , string memory gender, string memory hospital_name) public {
+    function createDoctor(string memory name , uint256 age , string memory gender, string memory hospital_name) public {
         Doctor memory newDoc;
         newDoc.name = name;
         newDoc.wallet_address = msg.sender;
@@ -69,12 +69,12 @@ contract HealthChain {
         newDoc.gender = gender;
         newDoc.hospital_name = hospital_name;
         doctors[msg.sender] = newDoc;
-        person_roles[msg.sender] =role_docter;
+        person_roles[msg.sender] = role_docter;
     }
     function getAllPatients() public view returns (Patient[] memory) {
         Doctor memory curr_doctor = doctors[msg.sender];
         Patient[] memory doctorPatients = new Patient[](curr_doctor.patients.length);
-        for (uint i = 0; i < curr_doctor.patients.length; i++) {
+        for (uint256 i = 0; i < curr_doctor.patients.length; i++) {
             doctorPatients[i] = patients[curr_doctor.patients[i]];
         }
         return doctorPatients;
@@ -85,7 +85,7 @@ contract HealthChain {
     }
 
     // Patient functions
-    function createPatient(string memory name, uint age, string memory gender) public {
+    function createPatient(string memory name, uint256 age, string memory gender) public {
         Patient memory newPatient;
         newPatient.name = name;
         newPatient.age = age;
@@ -102,7 +102,7 @@ contract HealthChain {
     function getIllnesses() public view returns (IllnessRes[] memory) {
         Illness[] memory curr_ills = patients_illness_map[msg.sender];
         IllnessRes[] memory ans = new IllnessRes[](curr_ills.length);
-        for(uint i=0;i<curr_ills.length;i++){
+        for(uint256 i=0;i<curr_ills.length;i++){
             ans[i].ill_id = curr_ills[i].ill_id;
             ans[i].name = curr_ills[i].name;
             ans[i].created_at = curr_ills[i].created_at;
@@ -121,7 +121,7 @@ contract HealthChain {
         Doctor memory doctor_to_add = doctors[doctor_address];
         address[] memory old_patients = doctor_to_add.patients;
         address[] memory new_patients = new address[](old_patients.length+1);
-        for(uint i =0;i<old_patients.length;i++){
+        for(uint256 i =0;i<old_patients.length;i++){
             new_patients[i]  = old_patients[i];
         }
         new_patients[old_patients.length] = msg.sender;
@@ -130,7 +130,7 @@ contract HealthChain {
         illnuses[ill_id] = newIllness;
         Illness[] memory curr_illnuses = patients_illness_map[msg.sender];
         Illness[] memory new_ills = new Illness[](curr_illnuses.length+1);
-        for(uint i =0;i<curr_illnuses.length;i++){
+        for(uint256 i =0;i<curr_illnuses.length;i++){
             new_ills[i] = curr_illnuses[i];
         }
         new_ills[curr_illnuses.length] = newIllness;
@@ -142,7 +142,7 @@ contract HealthChain {
         newRec.upload_date = upload_date;
         Record[] memory curr_reports = illness_record_map[ill_id];
         Record[] memory new_reports = new Record[](curr_reports.length+1);
-        for(uint i =0;i<curr_reports.length;i++){
+        for(uint256 i =0;i<curr_reports.length;i++){
             new_reports[i] = curr_reports[i];
         }
         new_reports[curr_reports.length] = newRec;
